# Conflicts:
#	src/main/java/Sys/Tick.java

+++ src/main/java/Sys/CalorieCheck.java
@@ -42,8 +42,8 @@ public class CalorieCheck implements Mode{
     /**
      *
      */
-    private int Calorie;
-    public int getCalorie() {
+    private double Calorie;
+    public double getCalorie() {
         calculateCalorie();
         return Calorie;
     }
@@ -186,7 +186,7 @@ public class CalorieCheck implements Mode{
     }
 
     private void calculateCalorie(){
-        int allSeconds = CalorieTime.getHour()*3600 + CalorieTime.getMinute()*60
+        double allSeconds = CalorieTime.getHour()*3600 + CalorieTime.getMinute()*60
                 + CalorieTime.getSecond();
         Calorie = (int) (0.0157 * ( ( 0.1 * Speed + 3.5 ) /3.5 ) * Weight * allSeconds);
     }
